requirements-scenario.txt
# Partnership-Enabled Scenario System - Python Dependencies
# For Task 67: Scenario YAML Schema and Validation

# YAML parsing
PyYAML>=6.0.1

# JSON Schema validation
jsonschema>=4.19.0

# CLI framework (for Task 69)
typer[all]>=0.9.0

# HTTP client for Research MCP (for Task 73)
httpx>=0.25.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Code quality
black>=23.7.0
ruff>=0.0.285
mypy>=1.5.0

